Refactor to golang
